% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_name_from.R
\name{add_name_from}
\alias{add_name_from}
\alias{add_city_from}
\title{Extract a pattern from a string in a data frame column}
\usage{
add_name_from(data, col, pattern = "^[^,]+")

add_city_from(data, col, pattern = "[^,]+$")
}
\arguments{
\item{data}{A data frame.}

\item{col}{The name of a column, e.g. "company_name".}

\item{pattern}{A pattern to extract from \code{col}.}
}
\value{
A data frame with a new column containing the extracted text.
}
\description{
Extract a pattern from a string in a data frame column
}
\examples{
library(tibble)

data <- tibble(company_name = "Peter GmbH & Co. KG, Berlin")
data %>%
  add_name_from("company_name") %>%
  add_city_from("company_name")

# Assumes that 'name' is after the last comma. If not the output is wrong.
data <- tibble(company_name = "bad, name")
data %>%
  add_name_from("company_name")

# Assumes that 'city' is after the last comma. If not the output is wrong.
data <- tibble(company_name = "city, bad")
data %>%
  add_city_from("company_name")
}
