\name{bundesbank-package}
\alias{bundesbank-package}
\alias{bundesbank}
\docType{package}
\title{
Download Data from Bundesbank Website
}
\description{
A helper function to download data from the website of the Bundesbank.
}
\details{

  The package contains only one function, \code{\link{getSeries}}, which
  allows to download data from the online database of the Bundesbank,
  the German central bank.
  
}
\author{
  Enrico Schumann

  Maintainer: Enrico Schumann <es@enricoschumann.net>

}
