% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bundesbank-package.R
\docType{package}
\name{bundesbank-package}
\alias{bundesbank}
\alias{bundesbank-package}
\title{bundesbank: Helpers for Bundesbank}
\description{
Provides helpers for Bundesbank.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://2degreesinvesting.github.io/bundesbank/}
}

}
\author{
\strong{Maintainer}: Mauro Lepore \email{maurolepore@gmail.com} (\href{https://orcid.org/0000-0002-1986-7988}{ORCID})

}
\keyword{internal}
